# Perf: compute_soft_cross_entropy